+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 ### **Copyrigth**: 
 *Tiledesk SRL*
 
+
+# 1.25.0-rc.3
+- **added**: intent and chatbot only agents visibility
+
+# 1.25.0-rc.2
+- **changed**: action-record translations and default values
+
 # 1.25.1
 
 # 1.25.0